refactor: SegmentedRoot -> Segmented 네이밍 변경
