+++ src/shared/ui/segmented-control/index.ts
@@ -1,4 +1,4 @@
-export { default as SegmentedControlRoot } from './segmented-control-root'
+export { default as SegmentedControl } from './segmented-control'
 export { default as SegmentedControlList } from './segmented-control-list'
 export { default as SegmentedControlItem } from './segmented-control-item'
 export { default as SegmentedControlContent } from './segmented-control-content'
